Documentation and refactoring for code related to seats

+++ src/main/java/net/cinemaApplication/backend/service/implementations/SeatImplementation.java
@@ -1,7 +1,9 @@
 package net.cinemaApplication.backend.service.implementations;
 
+import net.cinemaApplication.backend.entity.cinemaHall.CinemaHall;
 import net.cinemaApplication.backend.entity.cinemaHall.Seat;
 import net.cinemaApplication.backend.repository.SeatRepository;
+import net.cinemaApplication.backend.service.services.CinemaHallService;
 import net.cinemaApplication.backend.service.services.SeatService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -9,28 +11,60 @@ import org.springframework.stereotype.Service;
 import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
+import java.util.Optional;
+
 @Service
 public class SeatImplementation implements SeatService {
     @Autowired
     private SeatRepository seatRepository;
+    @Autowired
+    private CinemaHallService cinemaHallService;
 
+    /**
+     *
+     * @return a list of all the seats in the system
+     */
     @Override
     public List<Seat> getAllSeats() {
         return seatRepository.findAll();
     }
 
-//    @Override //for updating seat status when someone purchases a ticket or cancels
-//    public Seat updateSeatStatus(Long id, boolean status) {
-//        Seat seatFromDb = seatRepository.findById(id).get();
-//        if (seatFromDb.isSeatTaken() && status) {
-//            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "This seat is already taken");
-//        }
-//        seatFromDb.setSeatTaken(status);
-//        return seatRepository.save(seatFromDb);
-//    }
+    /**
+     *
+     * @return a seat by id
+     */
+    @Override
+    public Optional<Seat> getSeatById(Long id) {
+        return seatRepository.findById(id);
+    }
 
+
+    /**
+     * Delete a seat by id, also deletes tickets related to this seat
+     * @param id of the seat
+     * @throws ResponseStatusException if seat is missing
+     */
     @Override
     public void deleteById(Long id) {
+        if (seatRepository.findById(id).isEmpty()) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Seat not found");
+        }
         seatRepository.deleteById(id);
+    }
+
+    /**
+     * Get all the seats from a specific cinema hall
+     * @param hallId represents a cinema hall where the seats are
+     * @return list of seats in this cinema hall
+     * @throws ResponseStatusException if cinema hall is missing
+     */
+    @Override
+    public List<Seat> getAllSeatsFromCinemaHall(Long hallId) {
+        Optional<CinemaHall> cinemaHallOptional = cinemaHallService.getCinemaHallById(hallId);
+        if (cinemaHallOptional.isEmpty()) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Cinema hall not found");
+        }
+        CinemaHall cinemaHall = cinemaHallOptional.get();
+        return getAllSeats().stream().filter(c -> c.getHall() == cinemaHall).toList();
     }
 }
